+++ sdk/servicebus/azure-servicebus/tests/async_tests/test_queues_async.py
@@ -25,11 +25,11 @@ from azure.servicebus.exceptions import (
     ServiceBusConnectionError,
     ServiceBusError,
     MessageLockExpired,
-    InvalidHandlerState,
     MessageAlreadySettled,
     AutoLockRenewTimeout,
     MessageSendFailed,
-    MessageSettleFailed)
+    MessageSettleFailed,
+    MessageContentTooLarge)
 from devtools_testutils import AzureMgmtTestCase, CachedResourceGroupPreparer
 from servicebus_preparer import CachedServiceBusNamespacePreparer, CachedServiceBusQueuePreparer, ServiceBusQueuePreparer
 from utilities import get_logger, print_message, sleep_until_expired
@@ -541,9 +541,6 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                         count += 1
                     messages = await receiver.receive()
 
-                with pytest.raises(InvalidHandlerState):
-                    await receiver.receive()
-
             assert count == 10
 
             async with await sb_client.get_deadletter_receiver(idle_timeout=5, mode=ReceiveSettleMode.PeekLock) as receiver:
@@ -727,11 +724,11 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
             too_large = "A" * 1024 * 256
             
             async with sb_client.get_queue_sender(servicebus_queue.name) as sender:
-                with pytest.raises(MessageSendFailed):
+                with pytest.raises(MessageContentTooLarge):
                     await sender.send(Message(too_large))
                     
                 half_too_large = "A" * int((1024 * 256) / 2)
-                with pytest.raises(ValueError):
+                with pytest.raises(MessageContentTooLarge):
                     await sender.send([Message(half_too_large), Message(half_too_large)])